Chain email alerts to a successful content update…
Previously we were immediately sending an email
alert when the content editor hit ‘Save & Publish’
This is problematic because the content could fail
to update and the email alert would have already
been sent.

This changes it so that the email alert is sent
asynchronously on successful completion of a
content update. If this content update fails, then
no email alert will be sent until that job
succeeds.